requirements.txt
yfinance==1.3.0
pandas==2.3.3
numpy==2.2.6
ta==0.11.0
scikit-learn==1.5.2
matplotlib==3.10.9
torch==2.11.0
transformers==4.38.2
sqlalchemy==2.0.29
fredapi==0.5.2
lxml==5.1.0
html5lib==1.1
pykrx==1.2.7
psycopg2-binary==2.9.9
oci==2.133.0
python-dotenv==1.0.1
